+++ backend_flask/src/app/decorators.py
@@ -1,5 +1,6 @@
 from functools import wraps
-from flask import request, jsonify
+
+from flask import jsonify, request
 
 
 def validate_json(fn):
